feat(version): record CI build provenance and show it in `ddev version`
An unreleased DDEV build only carried a commit hash, which doesn't survive
a rebase and says nothing about which PR or CI run produced it. On GitHub
Actions, `versionconstants.BuildSource` is now set via -ldflags from a new
Makefile variable that reads the GITHUB_* variables Actions sets in every
job, recording the run URL and PR number when triggered by a pull_request
event. `scripts/build-source.sh` builds that value; it's a standalone
script rather than inline $(shell ...) because Make balances every
parenthesis in a shell function's argument, including literal ones meant
for the shell.

`ddev version` now shows a build-source row, but only when DdevVersion
isn't a clean release tag, so an official release's output is unchanged.
IsUnreleasedDdevVersion moved from pkg/ddevapp to pkg/versionconstants so
both pkg/version and pkg/ddevapp can use it without an import cycle
(pkg/ddevapp already imports pkg/version).
