+++ Makefile
@@ -33,7 +33,7 @@ PKG := github.com/ddev/ddev
 SRC_DIRS := cmd pkg
 
 # Version variables to replace in build
-VERSION_VARIABLES ?= DdevVersion AmplitudeAPIKey
+VERSION_VARIABLES ?= DdevVersion AmplitudeAPIKey BuildSource
 
 # These variables will be used as the default unless overridden by the make
 DdevVersion ?= $(VERSION)
@@ -46,6 +46,14 @@ DdevVersion ?= $(VERSION)
 # This version-strategy uses git tags to set the version string
 # VERSION can be overridden on make commandline: make VERSION=0.9.1 push
 VERSION := $(shell git describe --tags --always --dirty)
+
+# BuildSource ties an unreleased `ddev version` build back to the CI run it
+# came from, using the GITHUB_* variables GitHub Actions sets in every job.
+# Empty for a local `make` build, where none of them are set. Written as a
+# standalone script rather than inline shell: $(shell ...) balances every
+# parenthesis in its argument, including literal ones meant for the shell, so
+# anything more than a one-liner belongs in its own file.
+BuildSource := $(shell ./scripts/build-source.sh)
 # Some things insist on having the version without the leading 'v', so provide a
 # $(NO_V_VERSION) without it.
 # no_v_version removes the front v, for Chocolatey mostly
